refactor - shapes as a separate file
